refactor(tour): rename api/tutorial/tutorialSteps to api/tutorial/steps
