+++ client/src/api/tutorial/steps/discard.ts
